+++ packages/components/organisms/f-checkout/src/store/checkout.module.js
@@ -19,6 +19,8 @@ export default {
     state: () => ({
         id: '',
         serviceType: '',
+        restaurantId: '',
+        basketTotal: 0,
         customer: {
             firstName: '',
             lastName: '',
@@ -172,10 +174,50 @@ export default {
 
             const { data } = await axios.get(url, config);
             const basketDetails = {
-                serviceType: data.ServiceType.toLowerCase()
+                serviceType: data.ServiceType.toLowerCase(),
+                restaurantId: data.RestaurantId,
+                basketTotal: data.BasketSummary.BasketTotals.Total
             };
 
             commit(UPDATE_BASKET_DETAILS, basketDetails);
+        },
+
+        /**
+         * Get the address details from the backend and update the state.
+         *
+         * @param {Object} context - Vuex context object, this is the standard first parameter for actions
+         * @param {Object} payload - Parameter with the different configurations for the request.
+         */
+        getAddress: async ({ commit, state }, {
+            url,
+            language,
+            timeout
+        }) => {
+            const authHeader = state.authToken && `Bearer ${state.authToken}`;
+            const config = {
+                headers: {
+                    'Content-Type': 'application/json',
+                    'Accept-Language': language,
+                    ...(state.isLoggedIn && {
+                        Authorization: authHeader
+                    })
+                },
+                timeout
+            };
+
+            const { data } = await axios.get(url, config);
+
+            // TODO: Implement logic to select best address. For now, select the first address
+            const [selectedAddress] = data.Addresses;
+
+            const addressDetails = {
+                line1: selectedAddress.Line1,
+                line2: selectedAddress.Line2,
+                city: selectedAddress.City,
+                postcode: selectedAddress.ZipCode
+            };
+
+            commit(UPDATE_FULFILMENT_ADDRESS, addressDetails);
         },
 
         setAuthToken: ({ commit }, authToken) => {