+++ packages/components/organisms/f-status-banner/src/tenants/en-IE.js
@@ -1,0 +1,4 @@
+export default {
+    locale: 'en-IE',
+    text: 'I am a StatusBanner Component (IE)'
+};