+++ packages/components/organisms/f-status-banner/src/tenants/es-ES.js
@@ -1,0 +1,4 @@
+export default {
+    locale: 'es-ES',
+    text: 'I am a StatusBanner Component (ES)'
+};